Adjust typos. Remove logs. Adjust code
