+++ app/init/locales/en.json
@@ -4,5 +4,7 @@
 	"internal_server_error": "There was a problem executing the request, please try again later.",
 	"validator_missing_fields": "Some fields are incorrect or missing in {{ type }} : {{ extra }}",
 	"general_server_error": "There was an internal issue, please try again later.",
-	"not_found_page": "Page Not Found"
+	"not_found_page": "Page Not Found",
+	"invalid_body": "Invalid Input, pleae check you are sending the origianl an mutation objects"
+
 }
