+++ app/init/locales/es.json
@@ -4,5 +4,6 @@
 	"internal_server_error": "Se produjo un problema al ejecutar la solicitud. Vuelve a intentarlo más tarde.",
 	"validator_missing_fields": "Algunos campos son incorrectos o hacen falta en {{ type }} : {{extra}} ",
 	"general_server_error":"Hubo un problema interno. Vuelve a intentarlo más tarde.",
-	"not_found_page": "Pagina no encontrada"
+	"not_found_page": "Pagina no encontrada",
+	"invalid_body": "Datos incorrecots, por favor verificar que se envia le original y las mutaciones"
 }
