+++ app/modules/statement/controllers/statement.controller.js
@@ -1,7 +1,7 @@
 const fileUtil = rootRequire('lib/utils/files.util')
 const i18n = require('i18n')
 
-const StamentService = fileUtil.requireService('stament')
+const StatementService = fileUtil.requireService('statement')
 
 /**
  * This method allow generate the database type that is passed
@@ -13,7 +13,12 @@ const StamentService = fileUtil.requireService('stament')
  */
 exports.generateUpdateStatement = async (req, res) => {
 	try {
-		const serviceInstance = new StamentService()
+		const serviceInstance = new StatementService()
+
+		if(!req.body.original || !req.body.mutation){
+			res.badRequest({userMessage:  i18n.__('invalid_body'), serverInfo: req.body  })
+		}
+
 		const result = await serviceInstance.generateUpdateStatement(req.body)
 		res.ok({ data: result })
 	} catch (e) {
