+++ app/modules/statement/test/index.test.js
@@ -7,13 +7,12 @@ const fixtures = require('./fixtures')
 const chai = require('chai')
 
 const fileUtil = rootRequire('lib/utils/files.util')
-const StamentService = fileUtil.requireService('stament')
+const StatementService = fileUtil.requireService('statement')
 
 const { expect } = chai
 
 describe('Example Test', () => {
 	before(() => {
-		this.serviceInstance = new StamentService()
 		this.reqBody = {
 			original: fixtures.original,
 		}
@@ -23,7 +22,7 @@ describe('Example Test', () => {
 		it('should add a route document', async () => {
 			this.reqBody.mutation = fixtures.singleAdd
 
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$add')
@@ -32,7 +31,7 @@ describe('Example Test', () => {
 
 		it('should add a document in a leaf', async () => {
 			this.reqBody.mutation = fixtures.addDeep
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$add')
@@ -43,14 +42,14 @@ describe('Example Test', () => {
 	describe('$UPDATE', () => {
 		it('should update a route document', async () => {
 			this.reqBody.mutation = fixtures.singleUpdate
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 			expect(response).to.have.property('$update')
 		})
 
 		it('should update a document in a leaf', async () => {
 			this.reqBody.mutation = fixtures.updateDeep
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$update')
@@ -58,7 +57,7 @@ describe('Example Test', () => {
 
 		it('should update route document and document in a leaf', async () => {
 			this.reqBody.mutation = fixtures.updateSingleAndLeaf
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$update')
@@ -68,7 +67,7 @@ describe('Example Test', () => {
 	describe('$REMOVE', () => {
 		it('should remove a route document', async () => {
 			this.reqBody.mutation = fixtures.singleDelete
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$remove')
@@ -76,7 +75,7 @@ describe('Example Test', () => {
 
 		it('should remove a document in a leaf', async () => {
 			this.reqBody.mutation = fixtures.deleteDeep
-			const serviceInstance = new StamentService()
+			const serviceInstance = new StatementService()
 			const response = await serviceInstance.generateUpdateStatement(this.reqBody)
 
 			expect(response).to.have.property('$remove')
